Basic support for Lunar Magic edited roms.
Mods that use ASM hooks will not work except for wall jump.

You might have to edit the ROM with the latest lunar magic (3.33) and
edit a level and save it, and edit a map16 tile and save it, just to
make sure Lunar Magic upgrades the code in the ROM to the latest version.

Run the asset tool like:
python3 assets/restool.py -f -r "enhanced/enhanced.sfc"